I have created the tf_genarator package and changed it to use StaticTransformBroadcaster instead of call.
