+++ ros/src/util/packages/runtime_manager/scripts/map.yaml
@@ -94,7 +94,7 @@ params :
       v         : ''
       cmd_param :
         delim     : ''
-        default   : $(rospack find runtime_manager)/scripts/tf_generate.launch
+        default   : $(rospack find tf_generator)/launch/tf_generate.launch
 
   - name  : pcd_filter
     vars  :
