--- ros/src/util/packages/runtime_manager/scripts/tf_generate.launch
@@ -1,3 +1,0 @@
-<launch>
-    <node pkg="runtime_manager" type="tf_generate.py" name="tf_generate" />
-</launch>
