+++ ros/src/util/packages/tf_generator/launch/tf_generate.launch
@@ -1,0 +1,3 @@
+<launch>
+    <node pkg="tf_generator" type="tf_generate.py" name="tf_generate" />
+</launch>
